45 add tenant support (#108)
Add, update form for tenants has been added.

+++ src/main/java/com/a5lab/tabr/domain/rings/RingRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RingRepository extends JpaRepository<Ring, Integer> {
+public interface RingRepository extends JpaRepository<Ring, Long> {
   Optional<Ring> findRingByTitle(String title);
 
 }
